requirements.txt
aiofiles==0.8.0
aioredis==2.0.1
aiosqlite==0.17.0
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
Babel==2.9.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.0
fastapi-admin==1.0.3
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
iso8601==0.1.16
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.0
orjson==3.6.7
passlib==1.7.4
pendulum==2.1.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pypika-tortoise==0.1.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tortoise-orm==0.18.1
typing-extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.2
zipp==3.7.0
